requirements.txt
fastapi==0.110.2
numpy==1.23.5
onnx==1.14.1
onnxruntime==1.16.3
Pillow==9.5.0
Pillow==10.3.0
pytest==8.1.1
torchvision==0.18.0+cu118
uvicorn==0.29.0
